# packages/contracts/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_validate_template_package_templates_validate_package_post": {
        "properties": {
          "package": {
            "contentMediaType": "application/octet-stream",
            "description": "Template package zip containing JSON and PNG/SVG assets",
            "title": "Package",
            "type": "string"
          }
        },
        "required": [
          "package"
        ],
        "title": "Body_validate_template_package_templates_validate_package_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TemplateCatalogItemResponse": {
        "properties": {
          "aliases": {
            "items": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "array"
          },
          "canvas_height": {
            "title": "Canvas Height",
            "type": "integer"
          },
          "canvas_width": {
            "title": "Canvas Width",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "readiness": {
            "$ref": "#/components/schemas/TemplateReadinessReport"
          },
          "safe_zone_summary": {
            "items": {
              "$ref": "#/components/schemas/TemplateSafeZoneSummaryResponse"
            },
            "title": "Safe Zone Summary",
            "type": "array"
          },
          "source": {
            "$ref": "#/components/schemas/TemplateSourceMetadata"
          },
          "supported_views": {
            "items": {
              "type": "string"
            },
            "title": "Supported Views",
            "type": "array"
          },
          "thumbnail_url": {
            "title": "Thumbnail Url",
            "type": "string"
          },
          "view": {
            "title": "View",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label",
          "view",
          "supported_views",
          "canvas_width",
          "canvas_height",
          "thumbnail_url",
          "source",
          "readiness",
          "safe_zone_summary"
        ],
        "title": "TemplateCatalogItemResponse",
        "type": "object"
      },
      "TemplateDetailResponse": {
        "properties": {
          "aliases": {
            "items": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "array"
          },
          "asset_slots": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Asset Slots",
            "type": "object"
          },
          "authorization": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authorization"
          },
          "canvas_height": {
            "title": "Canvas Height",
            "type": "integer"
          },
          "canvas_width": {
            "title": "Canvas Width",
            "type": "integer"
          },
          "dimensions": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dimensions"
          },
          "export_config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Export Config"
          },
          "forbidden_zones": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Forbidden Zones",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "readiness": {
            "$ref": "#/components/schemas/TemplateReadinessReport"
          },
          "safe_zone_summary": {
            "items": {
              "$ref": "#/components/schemas/TemplateSafeZoneSummaryResponse"
            },
            "title": "Safe Zone Summary",
            "type": "array"
          },
          "safe_zones": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Safe Zones",
            "type": "array"
          },
          "scale": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale"
          },
          "sections": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Sections",
            "type": "array"
          },
          "source": {
            "$ref": "#/components/schemas/TemplateSourceMetadata"
          },
          "supported_views": {
            "items": {
              "type": "string"
            },
            "title": "Supported Views",
            "type": "array"
          },
          "thumbnail_url": {
            "title": "Thumbnail Url",
            "type": "string"
          },
          "view": {
            "title": "View",
            "type": "string"
          },
          "view_assets": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "View Assets",
            "type": "object"
          }
        },
        "required": [
          "id",
          "label",
          "view",
          "supported_views",
          "canvas_width",
          "canvas_height",
          "thumbnail_url",
          "source",
          "readiness",
          "safe_zone_summary",
          "asset_slots",
          "safe_zones"
        ],
        "title": "TemplateDetailResponse",
        "type": "object"
      },
...
      "TemplatePackageValidationResponse": {
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "authorization": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authorization"
          },
          "files_checked": {
            "items": {
              "type": "string"
            },
            "title": "Files Checked",
            "type": "array"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/TemplatePackageValidationIssue"
            },
            "title": "Issues",
            "type": "array"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "source_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Class"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          }
        },
        "required": [
          "accepted"
        ],
        "title": "TemplatePackageValidationResponse",
        "type": "object"
      },
      "TemplateReadinessReport": {
        "properties": {
          "blocking_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Blocking Reasons",
            "type": "array"
          },
          "catalog_eligible": {
            "title": "Catalog Eligible",
            "type": "boolean"
          },
          "missing_asset_slots": {
            "items": {
              "enum": [
                "base",
                "body_mask",
                "window_mask",
                "wheel_mask",
                "handle_mask",
                "panel_lines",
                "safe_zones",
                "metadata",
                "thumbnail"
              ],
              "type": "string"
            },
            "title": "Missing Asset Slots",
            "type": "array"
          },
          "required_asset_slots": {
            "items": {
              "enum": [
                "base",
                "body_mask",
                "window_mask",
                "wheel_mask",
                "handle_mask",
                "panel_lines",
                "safe_zones",
                "metadata",
                "thumbnail"
              ],
              "type": "string"
            },
            "title": "Required Asset Slots",
            "type": "array"
          },
          "reusable_asset_allowed": {
            "title": "Reusable Asset Allowed",
            "type": "boolean"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "catalog_eligible",
          "reusable_asset_allowed"
        ],
        "title": "TemplateReadinessReport",
        "type": "object"
      },
      "TemplateSafeZoneSummaryResponse": {
        "properties": {
          "height": {
            "title": "Height",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "width": {
            "title": "Width",
            "type": "number"
          },
          "x": {
            "title": "X",
            "type": "number"
          },
          "y": {
            "title": "Y",
            "type": "number"
          }
        },
        "required": [
          "id",
          "label",
          "kind",
          "x",
          "y",
          "width",
          "height"
        ],
        "title": "TemplateSafeZoneSummaryResponse",
        "type": "object"
      },
      "TemplateSourceMetadata": {
        "properties": {
          "allowed_usage_scope": {
            "default": "",
            "title": "Allowed Usage Scope",
            "type": "string"
          },
          "audit_timestamp": {
            "title": "Audit Timestamp",
            "type": "string"
          },
          "distribution_allowed": {
            "default": false,
            "title": "Distribution Allowed",
            "type": "boolean"
          },
          "license_evidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Evidence"
          },
          "license_status": {
            "enum": [
              "approved",
              "conditional",
              "workspace_only",
              "reference_only",
              "blocked",
              "missing"
            ],
            "title": "License Status",
            "type": "string"
          },
          "rights_notes": {
            "default": "",
            "title": "Rights Notes",
            "type": "string"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "source_type": {
            "enum": [
              "internal_original",
              "licensed_template",
              "user_provided_with_rights",
              "third_party_reference_only",
              "web_crawled_image"
            ],
            "title": "Source Type",
            "type": "string"
          }
        },
        "required": [
          "source_type",
          "license_status",
          "audit_timestamp"
        ],
        "title": "TemplateSourceMetadata",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Control-plane API foundation for the carAgent workbench.",
    "title": "carAgent API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/templates": {
      "get": {
        "operationId": "list_templates_templates_get",
        "parameters": [
          {
            "in": "query",
            "name": "view",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 40,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "View"
            }
          },
          {
            "in": "query",
            "name": "catalog_eligible",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Catalog Eligible"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TemplateCatalogItemResponse"
                  },
                  "title": "Response List Templates Templates Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Templates",
        "tags": [
          "templates"
        ]
      }
    },
    "/templates/validate-package": {
      "post": {
        "operationId": "validate_template_package_templates_validate_package_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_validate_template_package_templates_validate_package_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplatePackageValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Template Package",
        "tags": [
          "templates"
        ]
      }
    },
    "/templates/{template_id}": {
      "get": {
        "operationId": "get_template_templates__template_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "template_id",
            "required": true,
            "schema": {
              "title": "Template Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Template",
        "tags": [
          "templates"
        ]
      }
    },
    "/templates/{template_id}/thumbnail.png": {
      "get": {
        "operationId": "get_template_thumbnail_templates__template_id__thumbnail_png_get",
        "parameters": [
          {
            "in": "path",
            "name": "template_id",
            "required": true,
            "schema": {
              "title": "Template Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Template Thumbnail",
        "tags": [
          "templates"
        ]
      }
    },
...
  }
}